Services and DB moved to class libraries
