+++ FromLocalsToLocals.Services/Services/EF/Interfaces/IPostsService.cs
@@ -1,9 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace FromLocalsToLocals.Web.Models.Services
+namespace FromLocalsToLocals.Services.EF
 {
     public interface IPostsService
     {
