+++ FromLocalsToLocals.Services/Services/EF/Interfaces/IReviewsService.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace FromLocalsToLocals.Web.Models.Services
+namespace FromLocalsToLocals.Services.EF
 {
     public interface IReviewsService
     {
