+++ FromLocalsToLocals.Services/Services/EF/ReviewsService.cs
@@ -1,14 +1,12 @@
 ﻿using FromLocalsToLocals.Contracts.Entities;
-using FromLocalsToLocals.Web.Database;
-using FromLocalsToLocals.Web.Models.Services;
+using FromLocalsToLocals.Database;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using static FromLocalsToLocals.Web.Utilities.TimeCalculator;
 
-namespace FromLocalsToLocals.Models.Services
+namespace FromLocalsToLocals.Services.EF
 {
     public class ReviewsService : IReviewsService
     {
