+++ FromLocalsToLocals.Utilities/FromLocalsToLocals.Utilities.csproj
@@ -8,6 +8,7 @@
     <PackageReference Include="Geocoding.net" Version="3.6.1" />
     <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="5.0.1" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    <PackageReference Include="SendGrid" Version="9.21.1" />
   </ItemGroup>
 
 </Project>
