+++ FromLocalsToLocals/Controllers/HomeController.cs
@@ -11,9 +11,10 @@ using Microsoft.AspNetCore.Localization;
 using Microsoft.AspNetCore.Identity;
 using FromLocalsToLocals.Contracts.Entities;
 using FromLocalsToLocals.Utilities.Helpers;
-using FromLocalsToLocals.Web.Models.Services;
-using FromLocalsToLocals.Web.Database;
+using FromLocalsToLocals.Services;
+using FromLocalsToLocals.Database;
 using FromLocalsToLocals.Web.Models.ViewModels;
+using FromLocalsToLocals.Services.EF;
 
 namespace FromLocalsToLocals.Web.Controllers
 {
@@ -21,14 +22,14 @@ namespace FromLocalsToLocals.Web.Controllers
     {
         private readonly IToastNotification _toastNotification;
 
-        private readonly IVendorServiceEF _vendorService;
+        private readonly IVendorService _vendorService;
 
         private readonly IStringLocalizer<HomeController> _localizer;
         private readonly UserManager<AppUser> _userManager;
         private readonly AppDbContext _context;
 
 
-        public HomeController(AppDbContext context, IStringLocalizer<HomeController> localizer, IVendorServiceEF vendorService, IToastNotification toastNotification, UserManager<AppUser> userManager )
+        public HomeController(AppDbContext context, IStringLocalizer<HomeController> localizer, IVendorService vendorService, IToastNotification toastNotification, UserManager<AppUser> userManager )
         {
             _localizer = localizer;
             _vendorService = vendorService;
