+++ FromLocalsToLocalsWEB.sln
@@ -11,7 +11,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FromLocalsToLocals.Contract
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FromLocalsToLocals.Services", "FromLocalsToLocals.Services\FromLocalsToLocals.Services.csproj", "{237F99CC-0AD1-4990-8FE6-18A30276746F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FromLocalsToLocals.Utilities", "FromLocalsToLocals.Utilities\FromLocalsToLocals.Utilities.csproj", "{E3DF913E-D4DC-4C5A-A20A-A8D5C2121FAB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FromLocalsToLocals.Utilities", "FromLocalsToLocals.Utilities\FromLocalsToLocals.Utilities.csproj", "{E3DF913E-D4DC-4C5A-A20A-A8D5C2121FAB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FromLocalsToLocals.Database", "FromLocalsToLocals.Database\FromLocalsToLocals.Database.csproj", "{AD9D7BF4-E4A2-4A2D-90EE-6269DDF84008}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,6 +41,10 @@ Global
 		{E3DF913E-D4DC-4C5A-A20A-A8D5C2121FAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E3DF913E-D4DC-4C5A-A20A-A8D5C2121FAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E3DF913E-D4DC-4C5A-A20A-A8D5C2121FAB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD9D7BF4-E4A2-4A2D-90EE-6269DDF84008}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD9D7BF4-E4A2-4A2D-90EE-6269DDF84008}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD9D7BF4-E4A2-4A2D-90EE-6269DDF84008}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD9D7BF4-E4A2-4A2D-90EE-6269DDF84008}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
